Break apart sales engine class.
