+++ lib/sales_engine.rb
@@ -1,8 +1,15 @@
 require_relative 'item_repository'
 require_relative 'merchant_repository'
 require_relative 'invoice_repository'
+require_relative 'se_merchants_methods'
+require_relative 'se_items_methods'
+require_relative 'se_invoices_methods'
 
 class SalesEngine
+  include SeMerchantsMethods
+  include SeItemsMethods
+  include SeInvoicesMethods
+
   attr_reader :files,
               :items,
               :merchants,
@@ -34,61 +41,5 @@ class SalesEngine
     if csv_hash[:invoices]
       InvoiceRepository.new(csv_hash[:invoices], self)
     end
-  end
-
-  def all_items
-    @items.all
-  end
-
-  def all_merchant_ids
-    @merchants.all_ids
-  end
-
-  def all_invoices_count
-    @invoices.all.count
-  end
-
-  def get_merchants_total_inventory
-    @merchants.get_total_inventory
-  end
-
-  def get_item_prices_by_merchant_id(merchant_id)
-    @items.get_item_prices_by_merchant_id(merchant_id)
-  end
-
-  def return_merchants_by_item_count_greater_than(outliers)
-    @merchants.return_merchants_by_item_count_greater_than(outliers)
-  end
-
-  def merchant_items_count(merchant_id)
-    @items.find_all_by_merchant_id(merchant_id).length
-  end
-
-  def find_all_by_merchant_id(merchant_id)
-    @items.find_all_by_merchant_id(merchant_id)
-  end
-
-  def get_merchant_by_id(merchant_id)
-    @merchants.find_by_id(merchant_id)
-  end
-
-  def average_invoices_per_merchant
-    @invoices.average_invoices_per_merchant
-  end
-
-  def get_count_by_merchant_id
-    @invoices.get_count_by_merchant_id
-  end
-
-  def get_number_of_invoices_by_day
-    @invoices.get_number_of_invoices_by_day
-  end
-
-  def get_number_of_invoices_by_day_of_the_week
-    @invoices.get_invoice_count_by_day_of_week
-  end
-
-  def number_of_invoices_by_status
-    @invoices.number_of_invoices_by_status
   end
 end
